requirements.txt
praw
openai
tiktoken
langchain
python-dotenv